# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.AddMemberToTeamRequest:
    properties:
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    required:
    - user_id
    type: object
...
  dtos.ChangeTeamMemberRoleRequest:
    properties:
      role:
        enum:
        - owner
        - admin
        - member
        - viewer
        example: admin
        type: string
    required:
    - role
    type: object
...
  dtos.CreateTeamRequest:
    properties:
      name:
        example: Frontend Team
        maxLength: 80
        minLength: 3
        type: string
    required:
    - name
    type: object
...
  dtos.UpdateTeamRequest:
    properties:
      name:
        example: Front Team
        maxLength: 80
        minLength: 3
        type: string
    required:
    - name
    type: object
  errors.BasicError:
    properties:
      error:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
...
  models.Team:
    properties:
      created_at:
        type: string
      creator_id:
        type: string
      id:
        type: string
      name:
        maxLength: 100
        minLength: 5
        type: string
      org_id:
        type: string
    required:
    - name
    - org_id
    type: object
  models.TeamMember:
    properties:
      id:
        type: string
      joined_at:
        type: string
      role:
        enum:
        - owner
        - admin
        - member
        - viewer
        type: string
      team_id:
        type: string
      user_id:
        type: string
    required:
    - role
    - team_id
    - user_id
    type: object
  models.TeamMemberResult:
    properties:
      image_url:
        type: string
      name:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  models.TeamWithMemberCount:
    properties:
      id:
        type: string
      membersCount:
        type: integer
      name:
        type: string
    type: object
...
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  utils.SuccessResponseWithTotal:
    properties:
      data: {}
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
      total:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: support@taskio.com
    name: API Support
  description: Task management system API with real-time features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Taskio API
  version: "1.0"
paths:
...
  /orgs/{id}/teams:
    get:
      consumes:
      - application/json
      description: Retrieves all teams within an organization
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Teams fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.TeamWithMemberCount'
                  type: array
              type: object
        "400":
          description: Invalid organization ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Not a member
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: List organization teams
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: Creates a new team within an organization (owner or admin only)
      parameters:
      - description: Organization ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Team details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateTeamRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Team created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Team'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Create team
      tags:
      - Teams
...
  /teams/{teamId}:
    patch:
      consumes:
      - application/json
      description: Updates a team's name (owner only)
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      - description: New team name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Team updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Team'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Only owner can update
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Update team name
      tags:
      - Teams
  /teams/{teamId}/members:
    get:
      consumes:
      - application/json
      description: Retrieves all members of a specific team
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Team members fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.TeamMemberResult'
                  type: array
              type: object
        "400":
          description: Invalid team ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Not a team member
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get team members
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: Adds a user to a team (owner or admin only)
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      - description: User ID to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.AddMemberToTeamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Member added successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.TeamMember'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Add member to team
      tags:
      - Teams
  /teams/{teamId}/members/{userId}:
    delete:
      consumes:
      - application/json
      description: Removes a user from a team (owner or admin only)
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      - description: User ID (UUID) to remove
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Member deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Remove member from team
      tags:
      - Teams
    put:
      consumes:
      - application/json
      description: Changes the role of a team member (owner or admin only)
      parameters:
      - description: Team ID (UUID)
        in: path
        name: teamId
        required: true
        type: string
      - description: User ID (UUID)
        in: path
        name: userId
        required: true
        type: string
      - description: New role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.ChangeTeamMemberRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role changed successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "403":
          description: Insufficient permissions
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Change team member role
      tags:
      - Teams
...
securityDefinitions:
  BearerAuth:
    description: JWT token stored in cookie
    in: cookie
    name: taskio-cookie
    type: apiKey
swagger: "2.0"